Install webpack and integrate app for its usage
