--- assets/index.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/extensions */
-
-import GameLogic from './gamelogic.js';
-
-
-GameLogic.init();
