+++ webpack.config.js
@@ -1,0 +1,27 @@
+/* eslint-disable linebreak-style */
+const path = require('path');
+
+module.exports = {
+  entry: './src/assets/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          'css-loader',
+        ],
+      },
+      {
+        test: /\.(png|svg|jpg|gif|mp3)$/,
+        use: [
+          'file-loader',
+        ],
+      },
+    ],
+  },
+};
